+++ pkg/ml/train/accgradients_test.go
@@ -39,7 +39,8 @@ func TestTrainer_AccumulateGradients(t *testing.T) {
 	numTrainerMetrics := len(trainer.TrainMetrics())
 	require.NoError(t, err)
 	for ii := range numSteps {
-		metrics := trainer.TrainStep(nil, []*tensors.Tensor{input}, []*tensors.Tensor{label})
+		metrics, err := trainer.TrainStep(nil, []*tensors.Tensor{input}, []*tensors.Tensor{label})
+		require.NoError(t, err)
 		require.Len(t, metrics, numTrainerMetrics)
 
 		// Since the gradient hasn't been applied yet, the loss should be 10.0.